note area mode match
So now, the note area is same color of main text area, which makes sense.
Not sure if I like it though, maybe just a matter of getting used to it.
Maybe note area should always be white? Will leave as is for now.
Also, put 2 lines between each function, read somewhere that's how it should be.